Move compareHashes to util
